refactor: 변경사항에 맞게 스토리북 수정
